Combine wire.ts and contract.ts into decoders.ts
