--- packages/decoder/lib/decoders/index.ts
@@ -1,5 +1,0 @@
-import ContractDecoder from "./contract";
-import { ContractInstanceDecoder } from "./contract";
-import WireDecoder from "./wire";
-
-export { ContractDecoder, ContractInstanceDecoder, WireDecoder };
